refactor(takserver): commonize TAK SDK pipeline, drop redundant zstd/xpp3 deps
takpacket-sdk 0.7.0 (#5844) is a full KMP migration: its entire
conversion/compression pipeline now lives in the SDK's commonMain, zstd
runs on the transitive pure-Kotlin kzstd codec, and CoT XML on transitive
xmlutil. That makes the app's JVM-only scaffolding obsolete.

- Drop the explicit com.github.luben:zstd-jni (x3) and org.ogce:xpp3
  declarations plus both exclude(...) rules in core/model & core/takserver.
  Verified vestigial: the 0.7.0 SDK bytecode references kzstd (not
  com.github.luben) and xmlutil (not xpp3), and kzstd has zero runtime deps.
- Move the takpacket-sdk dependency from jvmAndroidMain to commonMain (api)
  in core/model.
- Collapse the TakV2Compressor / TakSdkCompressor / CotSanitizer
  expect/actual seams into single concrete commonMain objects calling the
  SDK directly; delete the 6 jvmAndroidMain + iosMain actual files. The old
  native-.so <clinit> guard (@Synchronized/@Volatile lazy + "zstd-jni
  unavailable" catch) is replaced by a plain `by lazy` — there is no native
  library left to fail.

iOS gains real zstd-dictionary compression and the golden-tested sanitize
rules in place of the previous passthrough/uncompressed-only/throwing stubs.
Transport/filesystem/TLS seams (TAKServer, AtakFileWriter, ZipArchiver,
TakCertLoader, TakFixtureLoader) remain platform-specific as before.
