+++ core/takserver/build.gradle.kts
@@ -37,11 +37,12 @@ kotlin {
             // org.meshtastic.proto.TAKPacketV2 and friends come from the
             // protobufs SDK (api()-exported by :core:model for every target).
             // The TAKPacket-SDK conversion pipeline (org.meshtastic.tak.*) is
-            // JVM-only since 0.5.2 and is api()-exported by :core:model's
-            // jvmAndroidMain — common code reaches it through the
-            // TakSdkCompressor/TakV2Compressor/CotSanitizer expect/actual seams.
-            // zstd-jni and xpp3 are excluded at the dep site in :core:model
-            // and re-added per-target below.
+            // multiplatform since 0.7.0 but consumed only on JVM/Android — it's
+            // api()-exported by :core:model's jvmAndroidMain, and common code
+            // reaches it through the TakSdkCompressor/TakV2Compressor/CotSanitizer
+            // expect/actual seams (iOS keeps stub actuals). zstd compression and
+            // CoT XML now ride on the SDK's transitive kzstd + xmlutil, so there
+            // are no native zstd-jni/xpp3 deps to re-add per target.
 
             implementation(libs.okio)
             implementation(libs.kotlinx.serialization.json)
@@ -54,37 +55,12 @@ kotlin {
             implementation(libs.kermit)
         }
 
-        jvmAndroidMain.dependencies {}
-
-        jvmMain.dependencies {
-            // Desktop JVM: standard JAR bundles native libs for desktop archs.
-            implementation("com.github.luben:zstd-jni:1.5.7-11")
-            // xpp3 is excluded from jvmAndroidMain (Android ships it as a
-            // platform class), but Desktop JVM still needs it for XmlPullParser.
-            implementation("org.ogce:xpp3:1.1.6")
-        }
-
-        androidMain.dependencies {
-            // Android: @aar variant ships .so files for arm/arm64/x86/x86_64.
-            // Without this, zstd-jni's ZstdDictCompress.<clinit> throws
-            // UnsatisfiedLinkError and poisons TakV2Compressor permanently.
-            implementation("com.github.luben:zstd-jni:1.5.7-11@aar")
-        }
-
         commonTest.dependencies {
             implementation(projects.core.testing)
             implementation(libs.kotlinx.coroutines.test)
             implementation(libs.turbine)
             implementation(libs.kotest.assertions)
             implementation(libs.kotest.property)
-        }
-
-        val androidHostTest by getting {
-            dependencies {
-                // Host-JVM tests need the platform JAR (not AAR) for zstd native
-                // libs — the @aar from androidMain only ships ARM/x86 .so files.
-                implementation("com.github.luben:zstd-jni:1.5.7-11")
-            }
         }
     }
 }
